Remove console.log from Brocfile and disable jscs in Grunt watch again
My editor always tells me the file was changed and I should reload when jscs is run on the code. This is because we have jscs automatically fix the detected style problems - even if there aren't any. I need to investigate if it is possible to stop jscs from changing the files when nothing needs to be fixed.