Start separating picker's core code from the $.fn plugin
We have had semantic issues because of the previous design, namely that the using the picker with 2 input fields would make no sense given that the plugin was called on a jQuery collection completely independent of these inputs.
Also, we now consider functionality responsible for establishing a connection between picker and outlet (i.e. input element) not to be in the domain of the picker. Therefore, this code needs to live in a plugin, where it can easily be disabled and replaced by custom code.