Make directional selection a plugin
If you have a look at the diff you might think this is a little exaggerated:
Pulling 4 LOC from an existing method and making it a 40+ line plugin.
In fact, though, it has only been possible to implement the feature in 4 lines before because of rather unclean design decisions.
In addition, every line removed from the core code is one line less in the core-version of the library, making it more attractive to users who don't need the feature.