Switch from custom option description schema to json schema
I eventually want to be able to have descriptions and nice titles in the web demo. This alone makes it necessary to revise the previous schema-model. Also, I hoped I would be able to leverage existing work, most notably json-form. But unfortunately it turns out that this lib is unmaintained and broken for our concrete schema.